Add dependencies to store data in influxdb
* Objective

This commit objective is about adding the required dependencies to start
storing data in influxdb.

* Why

We will use influxdb as the backend for the FE events coming from
the dApp. This MR introduces the dependencies to store such events.
Still, we need to map the information coming from the events properly.

* How

This commit adds the dependencies and the structure for the service to
start storing the events.