Store service worker events in influx and Postgres
* Objective

This MR objective is about storing the service worker events coming from
the front end in two different databases (for different purposes),
Postgres and influxdb.

* Why

We need to start collecting data, and this is the first commit that
enables such an operation, we now collect Solana web requests coming
from the dApps with this MR:

* How

This commit finalizes the influx repository with all the details and
creates the Postgres repository along with its DB models and migrations.